+++ Arduino/sketches/report_temp/temperature.h
@@ -1,7 +1,8 @@
 #ifndef TEMPERATURE_H
 #define TEMPERATURE_H
+#include <stdint.h>
 #include "statemachine.h"
 #include "sharedstate.h"
 
-StateMachineChoice initializeTemperature(unsigned long measureEvery, SharedState* sharedState);
+StateMachineChoice initializeTemperature(uint32_t measureEvery, SharedState* sharedState);
 #endif
